backend/requirements.txt
fastapi
uvicorn[standard]

torch
torchvision
numpy<2

pillow

ultralytics

pymongo
boto3

google-genai

python-multipart
python-dotenv
